Resolve a timing issue in the initial setup.
The initial setup (where the stop/start commands are sent) was done too fast
and was confusing the cricut (at least, that's what I'm assuming it was). I've
added a 170ms delay to give it time to process things, and this seems to
reliably fix the initialisation problems I was seeing.