Fix for timing issues in initial setup.
When initialising the USB connection we send a stop/start sequence very
quickly. This papears to confuse the device, leading it to stop responding to
commands sent via USB. This patch introduces a fix for this problem, by
inserting a delay before writing the stop and start commands. At the moment
some trial and error has suggested a delay of 170ms, but this  may need to be
changed with more testing.