Cleanup and finalise gcode support.
This implementation is at least somewhat clean and includes some reasonable
support for debugging output, a reasonably clean parser, and support for all
the stuff that I've seen in any of the test gcode files I've looked at.

Additional features that I'm going to implement are G3 and useful support for
the G90/91 commands (though I have yet to see any files that would exercise
the non-default case).