search $CWD and $HOME for .libcutter config file
Check if we've got one in the current working directory,
and if not, check in the user home directory.

So far, all supported cutters use the same file.

If we need to support different cutters in the future,
deal with that then, but now users won't need to pass
in a configfile path for every operation.