Added move_to prior to starting ellipse
In the render_ellipse_callback() the first point for the curve_to()
call is a control point, and not the current position.  Adding a
move_to() prior to this ensures the cutter is at that first point, and
prevents a cut-to when starting a new ellipse.  This change makes the
circle and ellipse svg test cases cut cleanly.